app/pyproject.toml
[tool.poetry]
name = "Phone Email and Places API"
version = "0.1.0"
description = "Validate and format input phone number and email and lookup places and country information"
authors = ["Anjum <anjum@sahl.solutions>"]

[tool.poetry.dependencies]
python = "^3.10.4"
phonenumberslite = "^8.12.46"
hiredis = "^2.0.0"
pydantic = {extras = ["email"], version = "^1.9.0"}
fastapi = "^0.75.1"
redis = "^4.2.2"
requests = "^2.27.1"
pytest-cov = "^3.0.0"
typer = {extras = ["all"], version = "^0.4.1"}
aiofiles = "^0.8.0"
aiohttp = "^3.8.1"
humanize = "^4.1.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
gunicorn = "^20.1.0"


[tool.poetry.dev-dependencies]
devtools = {extras = ["pygments"], version = "^0.8.0"}
black = "^22.3.0"
pytest = "^7.1.2"
ipython = "^8.3.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
